// src/api/be/schemas/apiResponseLetterDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { LetterDto } from './letterDto';

export interface ApiResponseLetterDto {
  success?: boolean;
  /** @nullable */
  message?: string | null;
  data?: LetterDto | null;
}