// src/api/be/schemas/letterDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { LetterAuthorDto } from './letterAuthorDto';

export interface LetterDto {
  id?: number;
  body?: string;
  author?: LetterAuthorDto;
  /** @nullable */
  arrivedAt?: string | null;
  createdAt?: string;
  liked?: boolean;
  saved?: boolean;
  likeCount?: number;
  read?: boolean;
  isMine?: boolean;
}