// src/api/generated/schemas/composeLetterDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * TripBite API
 * OpenAPI spec version: 1.0
 */
import type { LetterLocationDto } from './letterLocationDto';

export interface ComposeLetterDto {
  /**
     * @minLength 1
     * @maxLength 20
     */
  body: string;
  location: LetterLocationDto;
  isAnonymous?: boolean;
}